Refactor experiment table header context menu
