+++ webview/src/experiments/components/table/Table.test.tsx
@@ -15,7 +15,7 @@ import React from 'react'
 import tableDataFixture from 'dvc/src/test/fixtures/expShow/base/tableData'
 import { EXPERIMENT_WORKSPACE_ID } from 'dvc/src/cli/dvc/contract'
 import styles from './styles.module.scss'
-import { SortOrder } from './header/ContextMenuContent'
+import { SortOrder } from './header/util'
 import { ExperimentsTable } from '../Experiments'
 import { vsCodeApi } from '../../../shared/api'
 import {
